MQTT_Client/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
paho-mqtt>=1.6.1